Three int/unsigned argument types read off the MakeString instantiation
`MakeString<T>` mangles T, and the shipped linker map shows which T's fold:
`void*`, `unsigned int`, `unsigned long`, `long` all share 823b2008, while
`int` and enums share 82610090.  So the charge names the argument's signedness
directly.

  ChallengeSystemJobs.cpp 214   `(int)profiles.size()` -> `profiles.size()`
  SongLayout.cpp 186            `(int)it->mElements.size()` -> `.size()`
  OptionsPanel.cpp 66           `int res` -> `unsigned int res`
                                (XShowTokenRedemptionUI returns a DWORD)

The two casts are decompiler artifacts, not retail source: retail's
instantiation is the one you get by passing `size()` unmolested.

pinned objdiff-cli-B: name_check 42.705450% -> 42.714455% (+1024 bytes,
+3 complete functions, none lost).  `none` byte-identical at 43.730507% /
4973656 bytes; none_guard ??_C@ fingerprint unmoved.

TRIED AND REVERTED, with the measurement: HamDirector.cpp 2184
MILO_LOG -> MILO_NOTIFY.  The charge there is `?Notify@Debug@@QAAXPBD@Z`
against `??6TextStream@@QAAAAV0@PBD@Z`, which reads like the five sites landed
in 9556890e -- but those were Warn->Notify, two methods of the SAME class
reached through the same `TheDebug` global, and this one crosses TextStream to
Debug.  `TheDebug` is a `Debug`, `operator<<` is TextStream's, and the two
calls do not have the same instruction stream: LoadRoutineBuilderData fell
from 100 to 99.957085 at `none` (-932 bytes).  A `different_function` pair
being spelled by two macros in the same header does not make the macros
interchangeable.

+++ src/lazer/net_ham/ChallengeSystemJobs.cpp
@@ -211,7 +211,7 @@ GetChallengeBadgeCountsJob::GetChallengeBadgeCountsJob(
     : RCJob("leaderboards/getchallengebadgecounts/", callback) {
     MILO_LOG("***********************************\n");
     MILO_LOG("GetChallengeBadgeCountsJob::GetChallengeBadgeCountsJob()\n");
-    MILO_LOG(MakeString("   profiles.size()   = %d\n", (int)profiles.size()));
+    MILO_LOG(MakeString("   profiles.size()   = %d\n", profiles.size()));
     for (int i = 0; i < profiles.size(); i++) {
         MILO_LOG(MakeString(
             "   profiles[%d]       = '%s'\n", i, profiles[i]->GetOnlineID()->ToString()
